requirements.txt
click==8.1.7
markdown-it-py==3.0.0
markdown-pdf==1.0
mdurl==0.1.2
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pypandoc==1.12
PyYAML==6.0.1
